+++ lib/src/utils/get_suitable_calendar_general_date_widget.dart
@@ -1,3 +1,4 @@
+import 'package:clean_calendar/src/state/properties_state.dart';
 import 'package:clean_calendar/src/ui/grid_view_elements/calendar_general_date_widgets.dart';
 import 'package:flutter/material.dart';
 
@@ -5,7 +6,9 @@ Widget getSuitableCalendarGeneralDateWidget(
     {required bool enableDenseViewForDates,
     required bool enableDenseSplashForDates,
     required DateTime pageViewElementDate,
-    required DateTime pageViewMonthDate}) {
+    required DateTime pageViewMonthDate,
+    required CalendarPropertiesState
+        watchCalendarPropertiesStateProviderValue}) {
   if (enableDenseViewForDates && enableDenseSplashForDates) {
     return CalendarGeneralDenseSplashDate(
       pageViewElementDate: pageViewElementDate,