+++ lib/src/utils/logger.dart
@@ -1,0 +1,5 @@
+import 'dart:developer';
+
+show(dynamic _) {
+  log(_);
+}