+++ lib/src/utils/simulate_tap_on_an_offset.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/gestures.dart';
+
+simulateTapOnAnOffset(Offset offset) {
+  GestureBinding.instance.handlePointerEvent(PointerDownEvent(
+    position: offset,
+  ));
+  // await Future.delayed(Duration(milliseconds: 500));
+  GestureBinding.instance.handlePointerEvent(PointerUpEvent(
+    position: offset,
+  ));
+}