+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.2.0-dev.1
+
+* Removed weird logs from package.
+* Optimized calendar scrolling and loading time tremendously.
+* Disable and hide different types of dates through their properties.
+* Customise dates look and feel by changing its properties.
+* Single date selection with customized behaviour.
+* Access selected dates.
+* Provide pre-selected dates.
+* Provide start and end dates for calendar view.
+* Provide initial month view for calendar view.
+* Provide current date for calendar view.
+* Improved documentation.
+
 ## 0.0.1-dev.3
 
 * Now, showing only public methods and classes.